Added mechansims for checking application status using a scheduler
This mechanisms uses the service token which can be configured in OpenShift in such a way that the paas-manager can read all the application status from the project
A new variable openshift.token has been created. It is important to set a valid service token in order that the paas-manager is able to view the status of the applications
