+++ src/main/java/org/project/openbaton/nubomedia/paas/core/openshift/ProjectManager.java
@@ -33,7 +33,6 @@ import javax.annotation.PostConstruct;
 /**
  * Created by maa on 26.01.16.
  */
-//BETA!!
 @Service
 public class ProjectManager {
