--- src/main/java/org/project/openbaton/nubomedia/paas/messages/BuildingStatus.java
@@ -1,35 +1,0 @@
-/*
- * Copyright (c) 2015-2016 Fraunhofer FOKUS
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package org.project.openbaton.nubomedia.paas.messages;
-
-/**
- * Created by maa on 08.10.15.
- */
-public enum BuildingStatus {
-
-    CREATED,
-    INITIALIZING,
-    INITIALISED,
-    DUPLICATED,
-    BUILDING,
-    BUILD_OK, //internal state not visible from outside, used to redirect requests to deploymentManager
-    DEPLOYNG,
-    RUNNING,
-    FAILED,
-    PAAS_RESOURCE_MISSING
-
-}
