+++ src/main/java/org/project/openbaton/nubomedia/paas/utils/OpenshiftProperties.java
@@ -28,6 +28,8 @@ public class OpenshiftProperties {
 
     private String baseURL;
     private String domainName;
+    private String project;
+    private String token;
 
     public String getBaseURL() {
         return baseURL;
@@ -43,5 +45,20 @@ public class OpenshiftProperties {
 
     public void setDomainName(String domainName) {
         this.domainName = domainName;
+    }
+    public String getProject() {
+        return project;
+    }
+
+    public void setProject(String project) {
+        this.project = project;
+    }
+
+    public String getToken() {
+        return token;
+    }
+
+    public void setToken(String token) {
+        this.token = token;
     }
 }
