+++ src/main/resources/application.properties
@@ -16,6 +16,8 @@
 
 openshift.baseURL=https://localhost:8443
 openshift.domainName=example.com
+openshift.project=nubomedia
+openshift.token=eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJrdWJlcm5ldGVzL3NlcnZpY2VhY2NvdW50Iiwia3ViZXJuZXRlcy5pby9zZXJ2aWNlYWNjb3VudC9uYW1lc3BhY2UiOiJudWJvbWVkaWEiLCJrdWJlcm5ldGVzLmlvL3NlcnZpY2VhY2NvdW50L3NlY3JldC5uYW1lIjoicGFhcy10ZXN0LXRva2VuLTJqYjRuIiwia3ViZXJuZXRlcy5pby9zZXJ2aWNlYWNjb3VudC9zZXJ2aWNlLWFjY291bnQubmFtZSI6InBhYXMtdGVzdCIsImt1YmVybmV0ZXMuaW8vc2VydmljZWFjY291bnQvc2VydmljZS1hY2NvdW50LnVpZCI6ImFlNTNjMzc1LTI1NzgtMTFlNi1iYjNhLTAwMWE2NDhmOWNmNiIsInN1YiI6InN5c3RlbTpzZXJ2aWNlYWNjb3VudDpudWJvbWVkaWE6cGFhcy10ZXN0In0.WwdDcOyqgIVAOB3Vcnkg8iomOJzTrMICz_hxF87fP29lhQKdcyC0U563ZQsjV_kUtedWLNTQrucx3S7aT8ds8BtoQ9VFm9oX0iQB7M3Hi37h4NYe8eVeEhM_gkLflMonywzVoBzDhR8_rMcYrE5q2Xqo36SuKHVyrSR-JlmdPl57CyjySoh-0oL-EqVx0_3Tl327FKjmONf2Yl3egjZzVwxt28oTCZv6KI4Cm3oZC2mbfNaktLHrm911eGiWxi1LqUi_FjWdKDFIb-LHMz2q9G4RQ4N36YPojB9bLOyQp9w5Xgvkk9EEHrvNBk4E3IIRBZSijMoZ4emBHlMgSj43NQ
 paas.internalURL=http://localhost:8081
 paas.vnfmIP=localhost
 paas.vnfmPort=9000
